Add files via upload
The Pump is a app project that i decided to start to see if i could create an algorithm that would solve the solution for needing to find a fuel pump or Electric charging station while on the road. It would provide nearest pumps to you on the road from your location and starting with the cheapest price to offer. In addition my app would provide information about the gas station and whether other services are available in case of flat tire, car wash, bathroom, or need a bite to eat. Right now its just a project but i'm also on the way to completing it.